Simple unit tests discovery
1. Make all the unit tests discoverable with simple `nosetests tests`
   call;
2. Remove occurrences of `import *` in tests folder
3. Fix pep8 in tests folder
