+++ tests/scheduler/__init__.py
@@ -11,8 +11,3 @@
 # implied.
 # See the License for the specific language governing permissions and#
 # limitations under the License.
-
-
-from cursor import *
-from scheduler import *
-from task import *
